Debugging of Pipeline Build
For some reason, the pipeline is not detecting the changes in cdk.json and is creating some functions with incompatible versions, causing the build for Beta to fail. Attempting to remove the context cashe to rebuild these functions with the new required versions. Additonally, I manually changed the versions to hopefully also fix the error. We'll see what happens!