Changes in cdk.json
Needed to make the lambdaFunctionRuntime be set to python 3.9. What has been creating the runtime issues (from what I understand) is that the aws_s3_deployment.BucketDeployment function automatically creates a Lambda function required for our environment in its construct backend and the default is python 3.7. However, python 3.7 is no longer supported and the change in cdk.json *should* set the default runtime for the construct's Lambda function to python 3.9.