Updated Packages to Use AWS CDK V2
Had to update all of the packages to AWS CDK V2. With AWS CDK V1, we were running into the issue of aws_s3_deployment.BucketDeployment in the Visit stack (visit/__init__.py) deploying a Lambda function in the backend called CustomCDKDeployment and it's runtime was Python 3.7, an unsupported version. As it is impossible to change the runtime of this custom Lambda function that we have no control over, this was the deciding factor in updating all of the packages to AWS CDK V2.