Removed BasePathMapping
Had BasePathMapping in SharedAPIGateway stack to allow the referenced domain name to be associated with the new API being creating in the stack. Trying to remove this to see if we do not need to remap the variable for things to work as we're running into an error with having it included.