Resorted back to old code
Ended up removing the old stacks that were no longer needed as they were causing conflicting CNAME record errors, as the old stacks were utilizing the resources we needed to create. Therefore, the old code would build the environment as intended and build the resources required for the envionment to operate.