Moved create_rest_api()
Needed to move the function create_rest_api() from api_gateway/shared_api_gateway.py to dns/__init__.py (into the MakerspaceDns class). The reasoning behind it is because MakerspaceDnsRecords class needs to reference the Rest API which used to be created AFTER the function call in makerspace.py. Therefore, now that we have it being created earlier in the process in makerspace.py, we will have the required resources to move forward with deployment and in the correct order. Additionally, I have upgraded some functions to utilize AWS CDK V2 ways of creating the aws_certificatemanager.Certificate certificates (makes more sense if you see it in the code changes).